Set invalid password for app user
The behaviour of creating a new user changed. Nowadays the password hash is set
to `!` by default which prevents any login without setting a password.

Since we want the app user to login just with a public key and don't want a
password for sudo actions, we now set it to `*`.