Use wget
The vagrant environment doesn't have curl, and wget does the same thing.

It would also be nice to not use a hardcoded URL here and load from the current environment. But the environments vary too much to be worth writing a condition for it.
So non-prod environments will be out of date, but meh.