Add default_server declarations
This allows us to specify behaviour when an alternate hostname is requested (eg prod3).

Because some listen params can only be applied once per port, they are all moved to the default_server declarations.

Self-signed SSL cert won't be accepted for browsing, but that's ok because we just want to prevent access.