Redirect favicon.ico to favicon.png
This is requested a lot, and is being routed by Rails, using unnecessary resources. Our favicon is defined in the HTMLusing png which is a way better format. But I think it's some kind of client optimisation that looks ahead when preparing to navigate to a page before it has parsed (or perhaps even requested) the HTML. So we can add a hint here which says exactly what the HTML says: load favicon.png.