Revert "Handle download error"
I mistakenly thought that the /500 path would return a http success code, but was wrong. It only does that after the static .html file has been saved.
I think that it would be better to update the app to serve a 200 when generating the template. But it's probably not worth it.

But now, the script won't abort if there's a real error. (eg if puma failed to start or some other global app error). So I will add a different check next.