Save static cach for 404 as well
Despite the app being configured to use ErrorsController#not_found, the render method seems to consider the public/404.html a pre-cached result and serves that instead.

I think that's a good thing, it will save resources.
🤔 Hmm.. or maybe the reason it's not like this is because then the menu items and site banner could be out of date.

In that case, maybe we should be saving the 500.html and 404.html files with a cron job instead.