Limit connection timeout to 10 seconds
(curl's retry feature only activates once a HTTP error has been returned)

It seems the connection is delayed while waiting for a workers, but when workers are failing to start we're stuck in an infinite loop. I think nginx will wait 141 sec (proxy_read_timeout).
If it's not connecting after 10 seconds, we have a problem and need to know.