Increase file descriptor limit for fairfood processes
We recently hit the maximum due to high traffic. It seems the default maximum is severely outdated, so I've doubled it for now. We could probably increase in the future (https://serverfault.com/questions/48717/practical-maximum-open-file-descriptors-ulimit-n-for-a-high-volume-system).

The limit is loaded when the user session begins. Does restarting puma result in loading the new config? I wasn't able to test on staging, so I'm simply assuming it does for now.